# Fix list tightness.