According to the specification, blank lines in a block quote doesn't separate list items:
https://spec.commonmark.org/0.30/#example-320

Therefore, the following example should be tight:

```markdown
- > - a
  >
- b
```

The specification also say that link reference definitions can be children of list items when checking list tightness: https://spec.commonmark.org/0.30/#example-317

Therefore, the following example should be loose:

```
- [aaa]: /

  [bbb]: /
- b
```

This commit fixes those problems with the following strategy:

- Using source end position and start position of adjoining elements to check tightness. 
  This requires adjusting source end position of some block types to exclude trailing blank lines. This introduces an incompatible change to `sourcepos` property of `Node`. If this cannot be acceptable, I will add a field to `Node` to hold source position for list tightness.

- Delaying removal of link reference definitions until the entire document is parsed.

## Re: CONTRIBUTING.md

> Ensure that all tests pass (make test).

```
$ make test
...
723 tests passed, 0 failed.
```

> Use make lint to check for problems.

```
$ make lint
npm run lint

> commonmark@0.30.0 lint
> eslint .

```

> Ensure that performance has not been affected (make bench before and after the change).

`master`:

```
commonmark.js x 3,754 ops/sec ±0.96% (92 runs sampled)
showdown.js x 407 ops/sec ±1.36% (85 runs sampled)
marked.js x 1,652 ops/sec ±1.28% (90 runs sampled)
markdown-it x 3,680 ops/sec ±1.02% (90 runs sampled)
```

This branch:

```
commonmark.js x 3,756 ops/sec ±1.14% (91 runs sampled)
showdown.js x 422 ops/sec ±0.81% (88 runs sampled)
marked.js x 1,690 ops/sec ±1.03% (92 runs sampled)
markdown-it x 3,724 ops/sec ±0.99% (91 runs sampled)
```