@taku0 2023-02-10
`cmark` works well for the former example but not the latter example.