@taku0 2023-02-10
`cmark` fails to handle the following case while the `master` branch `commonmark.js` can handle it:

```markdown
- ***
  [aaa]: /

  bbb
- c
```