@taku0 2023-02-10
Is the following example tight or loose? The `master` branch handles it as loose while this branch handles it tight.

```markdown
- <pre>

- </pre>
- a
```