@taku0 2023-02-11
Updated to exlucde trailing blank lines from HTML blocks when checking tightness.

This behaviour seems inconsistent with the fenced code block:

`````markdown
- ```
  the following blank line is a part of fenced code block. Therefore, not
  separates list item:

- aaa

* <pre>The current implementation trims trailing blank lines from HTML blocks.
  Therefore, the following line is **not** a part of HTML block, so it
  separates the list items:

* The HTML block ends together with its parent (list item).
* </pre>
  The preceding line starts a new HTML block, it also ends together with its
  parent.

-     Indented code blocks are explicitly stated to exclude trailling blank
      lines.  Therefore, the following line is not a part of indentated code
      block and separetes the list items:
      
- aaa

`````

Maybe we should update the spec to specify whether each block types contain trailing blank lines or not.