@jgm 2023-02-11
Ah.  I hadn't taken into account of the fact that the blank line belongs to the raw HTML block (which it does according to the spec).

The spec says: "A list is [loose](https://spec.commonmark.org/0.30/#loose) if any of its constituent list items are separated by blank lines, or if any of its constituent list items directly contain two block-level elements with a blank line between them."  So, this case isn't loose after all, and I guess this is a bug in the behavior of the reference implementation.