@taku0 2023-02-12
Fixed HTML blocks to include trailing blank lines. The previous example is now tight.