@jgm 2023-08-07
Thank you very much for that lucid and detailed explanation.

I believe that the reason I didn't use the "compare end line to start line of next block" method was that the closing line positions for indented code and lists included the blank lines following them.  For example:

```
 % bin/commonmark --sourcepos
1. a
   - b
   - c


2. d
<ol data-sourcepos="1:1-6:4">
<li data-sourcepos="1:1-5:0">
<p data-sourcepos="1:4-1:4">a</p>
<ul data-sourcepos="2:4-5:0">
<li data-sourcepos="2:4-2:6">b</li>
<li data-sourcepos="3:4-5:0">c</li>
</ul>
</li>
<li data-sourcepos="6:1-6:4">
<p data-sourcepos="6:4-6:4">d</p>
</li>
</ol>
```
Note that the inner list goes from lines 2-5, and the next item starts on 6, so this woldu be tight as judged by the "next line" test.

You say that you've modified the finalizers to produce more accurate end positions, which would fix the problem.  I saw this code for indented code blocks but didn't notice it for lists -- is it there?

Anyway, as long as this problem is dealt with, I agree that your approach is much cleaner!

Have you checked benchmarks and pathological tests to make sure that there are no bad effects?  I imagine that handling reference links at the end by walking the whole document might be *slightly* less performant than handling them in the finalizers, but I'm guessing the difference is miniscule.